requirements.txt
Jinja2==2.7.1
MarkupSafe==0.18
PyYAML==3.10
facebook-sdk==0.4.0
psycopg2==2.5.1
web.py==0.37
wsgiref==0.1.2
